+++ example/hello_world/display_thread_id.dart
@@ -2,7 +2,5 @@ import 'dart:isolate';
 
 import 'package:squadron/squadron.dart';
 
-void displayThreadId(String name) {
-  Squadron.config(
-      '$name running in ${Isolate.current.runtimeType} ${Isolate.current.hashCode}');
-}
+String get workerId =>
+    '${Squadron.id} (${Isolate.current.runtimeType} ${Isolate.current.hashCode})';