+++ example/test/test_services.dart
@@ -11,100 +11,176 @@ import 'generated/test_services.activator.g.dart';
 part 'generated/test_services.worker.g.dart';
 
 @SquadronService()
-class TestService1 {
-  TestService1() {
-    // parameter-less constructor
-  }
+class TestParameterLess {
+  // parameter-less
+  TestParameterLess();
 }
 
 @SquadronService()
-class TestService2 {
-  TestService2(int arg1);
+class TestReqPositional {
+  // required positional
+  TestReqPositional(int arg1);
 }
 
 @SquadronService()
-class TestService3 {
-  TestService3([int arg1 = 0]);
+class TestOptNullPositional {
+  // optional nullable positional
+  TestOptNullPositional([int? arg1]);
 }
 
 @SquadronService()
-class TestService4 {
-  TestService4({int arg1 = 0});
+class TestOptDefPositional {
+  // optional non-nullable positional with default value
+  TestOptDefPositional([int arg1 = 0]);
 }
 
 @SquadronService()
-class TestService5 {
-  TestService5(int arg1, {int? arg2});
+class TestOptNullDefPositional {
+  // optional nullable positional with default value
+  TestOptNullDefPositional([int? arg1 = 0]);
 }
 
 @SquadronService()
-class TestService6 {
-  TestService6(int arg1, [int? arg2]);
-}
-
-@SquadronService()
-class TestService7 {
-  TestService7(this.arg1);
+class TestReqPositionalField {
+  // required positional
+  TestReqPositionalField(this.arg1);
 
   final int arg1;
 }
 
 @SquadronService()
-class TestService8 {
-  TestService8(this.arg1, [this.arg2]);
-
-  final int arg1;
-  final int? arg2;
-}
-
-@SquadronService()
-class TestService9 {
-  TestService9(this.arg1, {this.arg2});
-
-  final int arg1;
-  final int? arg2;
-}
-
-@SquadronService()
-class TestService10 {
-  TestService10([this.arg1]);
+class TestOptNullPositionalField {
+  // optional nullable positional
+  TestOptNullPositionalField([this.arg1]);
 
   final int? arg1;
 }
 
 @SquadronService()
-class TestService11 {
-  TestService11({this.arg1 = 0});
+class TestOptDefPositionalField {
+  // optional non-nullable positional with default value
+  TestOptDefPositionalField([this.arg1 = 0]);
 
   final int arg1;
 }
 
 @SquadronService()
-class TestService12 {
-  TestService12(this._arg1);
+class TestOptNullDefPositionalField {
+  // optional nullable positional with default value
+  TestOptNullDefPositionalField([this.arg1 = 0]);
+
+  final int? arg1;
+}
+
+@SquadronService()
+class TestReqPositionalPrivateField {
+  // required positional
+  TestReqPositionalPrivateField(this._arg1);
 
   final int _arg1;
-
-  @override
-  String toString() => '$_arg1';
 }
 
 @SquadronService()
-class TestService13 {
-  TestService13([this._arg1]);
+class TestOptNullPositionalPrivateField {
+  // optional nullable positional
+  TestOptNullPositionalPrivateField([this._arg1]);
 
   final int? _arg1;
-
-  @override
-  String toString() => '$_arg1';
 }
 
 @SquadronService()
-class TestService14 {
-  TestService14([this.__arg1]);
+class TestOptDefPositionalPrivateField {
+  // optional non-nullable positional with default value
+  TestOptDefPositionalPrivateField([this._arg1 = 0]);
 
-  final int? __arg1;
+  final int _arg1;
+}
 
-  @override
-  String toString() => __arg1?.toString() ?? '';
+@SquadronService()
+class TestOptNullDefPositionalPrivateField {
+  // optional nullable positional with default value
+  TestOptNullDefPositionalPrivateField([this._arg1 = 0]);
+
+  final int? _arg1;
+}
+
+@SquadronService()
+class TestOptNullNamed {
+  // optional nullable named
+  TestOptNullNamed({int? arg1});
+}
+
+@SquadronService()
+class TestOptDefNamed {
+  // optional non-nullable named with default value
+  TestOptDefNamed({int arg1 = 0});
+}
+
+@SquadronService()
+class TestOptNullDefNamed {
+  // optional nullable named with default value
+  TestOptNullDefNamed({int? arg1 = 0});
+}
+
+@SquadronService()
+class TestOptNullNamedField {
+  // optional nullable named
+  TestOptNullNamedField({this.arg1});
+
+  final int? arg1;
+}
+
+@SquadronService()
+class TestOptDefNamedField {
+  // optional non-nullable named with default value
+  TestOptDefNamedField({this.arg1 = 0});
+
+  final int arg1;
+}
+
+@SquadronService()
+class TestOptNullDefNamedField {
+  // optional nullable named with default value
+  TestOptNullDefNamedField({this.arg1 = 0});
+
+  final int? arg1;
+}
+
+@SquadronService()
+class TestOptNullNamedPrivateField {
+  // optional nullable named
+  TestOptNullNamedPrivateField({int? arg1}) : _arg1 = arg1;
+
+  final int? _arg1;
+}
+
+@SquadronService()
+class TestOptDefNamedPrivateField {
+  // optional non-nullable named with default value
+  TestOptDefNamedPrivateField({int arg1 = 0}) : _arg1 = arg1;
+
+  final int _arg1;
+}
+
+@SquadronService()
+class TestOptNullDefNamedPrivateField {
+  // optional nullable named with default value
+  TestOptNullDefNamedPrivateField({int? arg1 = 0}) : _arg1 = arg1;
+
+  final int? _arg1;
+}
+
+// mutable services should not be used via a worker pool
+@SquadronService(pool: false)
+class TestOptNullDefNamedPrivateNonFinalField {
+  // optional nullable named with default value
+  TestOptNullDefNamedPrivateNonFinalField({int? state = 0}) : _state = state;
+
+  int? _state;
+
+  @SquadronMethod()
+  FutureOr<int?> getState() => _state;
+
+  @SquadronMethod()
+  FutureOr<void> setState(int? state) => _state = state;
 }